refactor: simplify ZeebeDb factories
- `DefaultZeebeDbFactory` is only used in tests
- `ZeebeRocksDbFactory` can be constructed directly by passing config
